# architecture/openapi-fcom-curation.yaml
openapi: 3.0.0
info:
  title: FCOM Curation & Management API
  version: 1.0.0
  description: |
    Centralized REST API for managing FCOM rule files across multiple Unified Assurance (UA) server environments.
    Interfaces with each UA server's native REST API and SVN version control system.
servers:
  - url: https://fcom-curator.example.com/api/v1
    description: Production API
...
paths:
...
  /servers:
    get:
      tags:
        - Server Configuration
      summary: List all configured UA server instances
      security:
        - sessionCookie: []
      responses:
        '200':
          description: List of available servers
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    server_id:
                      type: string
                      example: "prod-ua-01"
                    server_name:
                      type: string
                      example: "Production UA Primary"
                    hostname:
                      type: string
                      example: "ua.prod.example.com"
                    environment:
                      type: string
                      enum: [dev, test, staging, prod]
                    svn_url:
                      type: string
                      example: "svn://ua.prod.example.com/fcom"
  /servers/{server_id}/switch:
    post:
      tags:
        - Server Configuration
      summary: Switch session context to a different server
      description: |
        Changes the active server for the current session.
        If unsaved edits exist, the caller must handle the warning and save/discard before switching.
        Supports both basic auth and certificate-based auth methods.
      security:
        - sessionCookie: []
      parameters:
        - name: server_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required:
                    - auth_type
                    - username
                    - password
                  properties:
                    auth_type:
                      type: string
                      enum: [basic]
                    username:
                      type: string
                    password:
                      type: string
                      format: password
                - type: object
                  required:
                    - auth_type
                    - cert_path
                    - key_path
                  properties:
                    auth_type:
                      type: string
                      enum: [certificate]
                    cert_path:
                      type: string
                    key_path:
                      type: string
                    ca_cert_path:
                      type: string
      responses:
        '200':
          description: Session context switched
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
                  user:
                    type: string
                  server_id:
                    type: string
        '401':
          description: Invalid credentials or certificate for target server
        '404':
          description: Server not found
...
  /files/{file_id}/read:
    get:
      tags:
        - File Editor
      summary: Read and parse a complete FCOM JSON file
      description: |
        Fetches the full JSON content from SVN, parses it, and returns structured data.
        Performs initial schema validation and returns any validation errors.
      security:
        - sessionCookie: []
      parameters:
        - name: file_id
          in: path
          required: true
          schema:
            type: string
        - name: revision
          in: query
          schema:
            type: string
          description: Optional SVN revision specifier (e.g., "HEAD", "1234")
      responses:
        '200':
          description: File content with metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  file_id:
                    type: string
                  path:
                    type: string
                  revision:
                    type: string
                  last_modified:
                    type: string
                    format: date-time
                  last_author:
                    type: string
                  etag:
                    type: string
                    description: Content hash for optimistic locking
                  content:
                    type: object
                    description: Parsed JSON content
                  validation_errors:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                        message:
                          type: string
                        severity:
                          type: string
                          enum: [error, warning]
        '404':
          description: File not found
  /files/{file_id}/save:
    post:
      tags:
        - File Editor
      summary: Validate and commit changes to a FCOM JSON file
      description: |
        Validates the provided JSON against the FCOM schema.
        If valid, commits the changes to SVN with a user-provided commit message.
        Uses optimistic locking (etag) to prevent lost updates.
      security:
        - sessionCookie: []
      parameters:
        - name: file_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
                - etag
                - commit_message
              properties:
                content:
                  type: object
                  description: Updated JSON content
                etag:
                  type: string
                  description: etag from the last read (for conflict detection)
                commit_message:
                  type: string
                  description: SVN commit message describing the changes
      responses:
        '200':
          description: File saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  file_id:
                    type: string
                  revision:
                    type: string
                  last_modified:
                    type: string
                    format: date-time
                  commit_id:
                    type: string
                  etag:
                    type: string
        '400':
          description: Validation errors in the JSON
          content:
            application/json:
              schema:
                type: object
                properties:
                  validation_errors:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                        message:
                          type: string
        '409':
          description: Conflict - file has been modified by another user (etag mismatch)
  /files/{file_id}/diff:
    get:
      tags:
        - Version Control
      summary: Get diff between working copy and HEAD (or two revisions)
      description: |
        Returns a unified diff showing changes between the current working copy
        and the last committed version, or between two specified SVN revisions.
      security:
        - sessionCookie: []
      parameters:
        - name: file_id
          in: path
          required: true
          schema:
            type: string
        - name: revision_a
          in: query
          schema:
            type: string
            default: "HEAD"
        - name: revision_b
          in: query
          schema:
            type: string
            default: "WORKING"
      responses:
        '200':
          description: Diff output
          content:
            application/json:
              schema:
                type: object
                properties:
                  file_id:
                    type: string
                  revision_a:
                    type: string
                  revision_b:
                    type: string
                  unified_diff:
                    type: string
                    description: Unified diff format (can be rendered as-is or parsed)
                  summary:
                    type: object
                    properties:
                      additions:
                        type: integer
                      deletions:
                        type: integer
  /files/{file_id}/history:
    get:
      tags:
        - Version Control
      summary: Get SVN commit history for a file
      description: |
        Returns a paginated list of commits affecting this file,
        ordered most-recent first.
      security:
        - sessionCookie: []
      parameters:
        - name: file_id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Commit history
          content:
            application/json:
              schema:
                type: object
                properties:
                  file_id:
                    type: string
                  total:
                    type: integer
                  commits:
                    type: array
                    items:
                      type: object
                      properties:
                        revision:
                          type: string
                        author:
                          type: string
                        date:
                          type: string
                          format: date-time
                        message:
                          type: string
...
  /files/{file_id}/test:
    post:
      tags:
        - Testing
      summary: Generate and send a test trap for a single event object
      description: |
        Calls FCOM2Test (or equivalent) to generate a trap for a specific event object.
        Sends the test trap to the UA server and returns the result.
      security:
        - sessionCookie: []
      parameters:
        - name: file_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - object_name
              properties:
                object_name:
                  type: string
                  description: "@objectName of the event to test"
      responses:
        '200':
          description: Test execution result
          content:
            application/json:
              schema:
                type: object
                properties:
                  test_id:
                    type: string
                  object_name:
                    type: string
                  status:
                    type: string
                    enum: [success, failure, pending]
                  output:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
  /files/{file_id}/test-all:
    post:
      tags:
        - Testing
      summary: Test all event objects in a file
      description: |
        Iterates through all objects in the file and runs a test for each.
        Returns aggregate results and per-object status.
      security:
        - sessionCookie: []
      parameters:
        - name: file_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Batch test results
          content:
            application/json:
              schema:
                type: object
                properties:
                  file_id:
                    type: string
                  test_id:
                    type: string
                  total_objects:
                    type: integer
                  passed:
                    type: integer
                  failed:
                    type: integer
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        object_name:
                          type: string
                        status:
                          type: string
                          enum: [success, failure, pending]
                        message:
                          type: string
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: FCOM_SESSION_ID
      description: HTTP-only session cookie set after successful login
...
security:
  - sessionCookie: []