# fcom-curator/backend/src/routes/fileEditor.ts
import { Router, Request, Response } from 'express';
import crypto from 'crypto';
import logger from '../utils/logger';

const router = Router();

/**
 * GET /api/v1/files/:file_id/read
 * Read and parse a complete FCOM JSON file
 */
router.get('/:file_id/read', (req: Request, res: Response) => {
  try {
    const { file_id } = req.params;
    const { revision = 'HEAD' } = req.query;

    logger.info(`Reading file: ${file_id}, revision: ${revision}`);

    // TODO: Call UA server REST API to fetch file from SVN
    // Mock response
    const mockContent = {
      objects: [
        {
          '@objectName': 'CISCO-ENVMON-FAN-FAILURE',
          description: 'Fan failure detected in Cisco device',
          certification: 'CERTIFIED',
          event: {
            EventType: 'DEVICE_EVENT',
            Severity: 'MAJOR',
            Summary: 'Fan failure',
          },
          trap: {
            variables: ['$v1', '$v2'],
          },
        },
      ],
    };

    const etag = crypto.createHash('md5').update(JSON.stringify(mockContent)).digest('hex');

    res.json({
      file_id: file_id,
      path: `/trap/cisco/${file_id}`,
      revision: revision as string,
      last_modified: new Date().toISOString(),
      last_author: 'api',
      etag: etag,
      content: mockContent,
      validation_errors: [],
    });
  } catch (error: any) {
    logger.error(`Error reading file: ${error.message}`);
    res.status(500).json({ error: 'Failed to read file' });
  }
});

/**
 * POST /api/v1/files/:file_id/save
 * Validate and commit changes to a FCOM JSON file
 */
router.post('/:file_id/save', (req: Request, res: Response) => {
  try {
    const { file_id } = req.params;
    const { content, etag, commit_message } = req.body;

    if (!content || !etag || !commit_message) {
      return res.status(400).json({ error: 'Missing content, etag, or commit_message' });
    }

    logger.info(`Saving file: ${file_id}, message: ${commit_message}`);

    // TODO: Validate against FCOM schema
    // TODO: Call UA server REST API to commit to SVN

    const newEtag = crypto.createHash('md5').update(JSON.stringify(content)).digest('hex');

    res.json({
      file_id: file_id,
      revision: 'HEAD',
      last_modified: new Date().toISOString(),
      commit_id: 'mock-commit-id',
      etag: newEtag,
    });
  } catch (error: any) {
    logger.error(`Error saving file: ${error.message}`);
    res.status(500).json({ error: 'Failed to save file' });
  }
});

/**
 * GET /api/v1/files/:file_id/diff
 * Get diff between working copy and HEAD (or two revisions)
 */
router.get('/:file_id/diff', (req: Request, res: Response) => {
  try {
    const { file_id } = req.params;
    const { revision_a = 'HEAD', revision_b = 'WORKING' } = req.query;

    logger.info(`Getting diff for file: ${file_id}`);

    // TODO: Call UA server REST API for diff
    res.json({
      file_id: file_id,
      revision_a: revision_a,
      revision_b: revision_b,
      unified_diff: '--- a/file\n+++ b/file\n@@ -1,3 +1,4 @@\n context\n-removed line\n+added line',
      summary: {
        additions: 1,
        deletions: 1,
      },
    });
  } catch (error: any) {
    logger.error(`Error getting diff: ${error.message}`);
    res.status(500).json({ error: 'Failed to get diff' });
  }
});

/**
 * GET /api/v1/files/:file_id/history
 * Get SVN commit history for a file
 */
router.get('/:file_id/history', (req: Request, res: Response) => {
  try {
    const { file_id } = req.params;
    const { limit = 20, offset = 0 } = req.query;

    logger.info(`Getting history for file: ${file_id}`);

    // TODO: Call UA server REST API for history
    res.json({
      file_id: file_id,
      total: 5,
      commits: [
        {
          revision: '1234',
          author: 'user1',
          date: new Date().toISOString(),
          message: 'Updated FCOM object definition',
        },
        {
          revision: '1233',
          author: 'user2',
          date: new Date(Date.now() - 86400000).toISOString(),
          message: 'Initial creation',
        },
      ],
    });
  } catch (error: any) {
    logger.error(`Error getting history: ${error.message}`);
    res.status(500).json({ error: 'Failed to get history' });
  }
});

/**
 * POST /api/v1/files/:file_id/test
 * Generate and send a test trap for a single event object
 */
router.post('/:file_id/test', (req: Request, res: Response) => {
  try {
    const { file_id } = req.params;
    const { object_name } = req.body;

    if (!object_name) {
      return res.status(400).json({ error: 'Missing object_name' });
    }

    logger.info(`Testing object ${object_name} in file ${file_id}`);

    // TODO: Call FCOM2Test utility
    res.json({
      test_id: 'test-001',
      object_name: object_name,
      status: 'success',
      output: 'Trap sent successfully',
      timestamp: new Date().toISOString(),
    });
  } catch (error: any) {
    logger.error(`Error testing object: ${error.message}`);
    res.status(500).json({ error: 'Failed to test object' });
  }
});

/**
 * POST /api/v1/files/:file_id/test-all
 * Test all event objects in a file
 */
router.post('/:file_id/test-all', (req: Request, res: Response) => {
  try {
    const { file_id } = req.params;

    logger.info(`Running batch test for file ${file_id}`);

    // TODO: Iterate through all objects and test
    res.json({
      file_id: file_id,
      test_id: 'batch-test-001',
      total_objects: 3,
      passed: 3,
      failed: 0,
      results: [
        { object_name: 'OBJ1', status: 'success', message: 'OK' },
        { object_name: 'OBJ2', status: 'success', message: 'OK' },
        { object_name: 'OBJ3', status: 'success', message: 'OK' },
      ],
    });
  } catch (error: any) {
    logger.error(`Error running batch test: ${error.message}`);
    res.status(500).json({ error: 'Failed to run batch test' });
  }
});

export default router;